+++ examples/js/sort/demo.js
@@ -1,6 +1,7 @@
 /* eslint max-len: 0 */
 import React from 'react';
 import SortTable from './sort-table';
+import MultiSortTable from './multi-sort-table';
 import DefaultSortTable from './default-sort-table';
 import ExternalSort from './manage-sort-external-table';
 import CustomSortTable from './custom-sort-table';
@@ -9,6 +10,7 @@ import ReusableCustomSortTable from './reusable-custom-sort-table';
 import SortHookTable from './sort-hook-table';
 import DisableSortIndicatorTable from './disable-sort-indicator-table';
 import CustomCaretSortTable from './custom-caret-sort-table';
+import ExternalMultiSort from './manage-multi-sort-external-table';
 
 class Demo extends React.Component {
   render() {
@@ -20,6 +22,15 @@ class Demo extends React.Component {
             <div className='panel-body'>
               <h5>Source in /examples/js/sort/sort-table.js</h5>
               <SortTable />
+            </div>
+          </div>
+        </div>
+        <div className='col-md-offset-1 col-md-8'>
+          <div className='panel panel-default'>
+            <div className='panel-heading'>Table Multi Sort Example</div>
+            <div className='panel-body'>
+              <h5>Source in /examples/js/sort/multi-sort-table.js</h5>
+              <MultiSortTable />
             </div>
           </div>
         </div>
@@ -38,6 +49,15 @@ class Demo extends React.Component {
             <div className='panel-body'>
               <h5>Source in /examples/js/sort/manage-sort-external-table.js</h5>
               <ExternalSort />
+            </div>
+          </div>
+        </div>
+        <div className='col-md-offset-1 col-md-8'>
+          <div className='panel panel-default'>
+            <div className='panel-heading'>Manage Multi Sorting Externally Example</div>
+            <div className='panel-body'>
+              <h5>Source in /examples/js/sort/manage-multi-sort-external-table.js</h5>
+              <ExternalMultiSort />
             </div>
           </div>
         </div>