+++ src/Filter.js
@@ -7,8 +7,12 @@ export class Filter extends EventEmitter {
     this.currentFilter = {};
   }
 
-  handleFilter(dataField, value, type) {
+  handleFilter(dataField, value, type, filterObj) {
     const filterType = type || Const.FILTER_TYPE.CUSTOM;
+
+    const props = {
+      cond: filterObj.condition // Only for select and text filter
+    };
 
     if (value !== null && typeof value === 'object') {
       // value of the filter is an object
@@ -21,14 +25,14 @@ export class Filter extends EventEmitter {
       }
       // if one of the object properties is undefined or empty, we remove the filter
       if (hasValue) {
-        this.currentFilter[dataField] = { value: value, type: filterType };
+        this.currentFilter[dataField] = { value: value, type: filterType, props };
       } else {
         delete this.currentFilter[dataField];
       }
     } else if (!value || value.trim() === '') {
       delete this.currentFilter[dataField];
     } else {
-      this.currentFilter[dataField] = { value: value.trim(), type: filterType };
+      this.currentFilter[dataField] = { value: value.trim(), type: filterType, props };
     }
     this.emit('onFilterChange', this.currentFilter);
   }